Moved test decorators
